+++ backend/homework/3/src/main/java/hospitalManagementSystem/Billing.java
@@ -1,4 +1,4 @@
-package HMS;
+package hospitalManagementSystem;
 
 public class Billing {
