+++ backend/homework/3/src/main/java/hospitalManagementSystem/BlueCrossBlueShield.java
@@ -1,4 +1,4 @@
-package HMS;
+package hospitalManagementSystem;
 
 public class BlueCrossBlueShield implements InsuranceBrand{
     public double computeMonthlyPremium(HealthInsurancePlan insurancePlan, int age, boolean smoking){
