+++ backend/homework/3/src/main/java/hospitalManagementSystem/BronzePlan.java
@@ -1,4 +1,4 @@
-package HMS;
+package hospitalManagementSystem;
 
 public class BronzePlan extends HealthInsurancePlan{
     public BronzePlan(){
