+++ backend/homework/3/src/main/java/hospitalManagementSystem/HealthInsurancePlan.java
@@ -1,4 +1,4 @@
-package HMS;
+package hospitalManagementSystem;
 
 public abstract class HealthInsurancePlan {
     private double coverage;
