+++ backend/homework/3/src/main/java/hospitalManagementSystem/Main.java
@@ -1,4 +1,4 @@
-package HMS;
+package hospitalManagementSystem;
 
 import org.slf4j.LoggerFactory;
