+++ backend/homework/3/src/main/java/hospitalManagementSystem/Nurse.java
@@ -1,4 +1,4 @@
-package HMS;
+package hospitalManagementSystem;
 
 public class Nurse extends Staff{
     private long nurseId;
