+++ backend/homework/3/src/main/java/hospitalManagementSystem/Patient.java
@@ -1,4 +1,4 @@
-package HMS;
+package hospitalManagementSystem;
 
 public class Patient extends User{
     private long patientId;
