+++ backend/homework/3/src/main/java/hospitalManagementSystem/SilverPlan.java
@@ -1,4 +1,4 @@
-package HMS;
+package hospitalManagementSystem;
 
 public class SilverPlan extends HealthInsurancePlan{
     public SilverPlan(){
