+++ backend/homework/3/src/main/java/hospitalManagementSystem/User.java
@@ -1,4 +1,4 @@
-package HMS;
+package hospitalManagementSystem;
 
 public class User {
     private long id;
